feat(fips): return an error when validating kerberos cfg (#42887)
* feat(fips): return an error when validating kerberos cfg

setting kerberos config options should return an error

* Update config.go

* Update config_nofips.go

* Update client_nofips.go
